+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 
+- Scanners: Distinguish thinking, thinking summary, and redacted thinking when rendering messages for scanners.
 - Scout View: Show tooltip for array or object values which have been truncated.
 - VS Code Integration: Fix regression displaying Scout View in VS Code.